requirements.txt
matplotlib==3.10.0
numpy==2.2.2
Pillow==11.1.0
